refactor(html/minifier): Reorganize minification steps
