+++ crates/swc_html_minifier/src/lib.rs
@@ -3,198 +3,25 @@
 use swc_html_ast::*;
 use swc_html_visit::{VisitMut, VisitMutWith};
 
-static BOOLEAN_ATTRIBUTES: &[&str] = &[
-    "allowfullscreen",
-    "async",
-    "autofocus",
-    "autoplay",
-    "checked",
-    "compact",
-    "controls",
-    "declare",
-    "default",
-    "defaultchecked",
-    "defaultmuted",
-    "defaultselected",
-    "defer",
-    "disabled",
-    "enabled",
-    "formnovalidate",
-    "hidden",
-    "indeterminate",
-    "inert",
-    "ismap",
-    "itemscope",
-    "loop",
-    "multiple",
-    "muted",
-    "nohref",
-    "noresize",
-    "noshade",
-    "novalidate",
-    "nowrap",
-    "open",
-    "pauseonexit",
-    "readonly",
-    "required",
-    "reversed",
-    "scoped",
-    "seamless",
-    "selected",
-    "sortable",
-    "truespeed",
-    "typemustmatch",
-    "visible",
-];
+mod steps;
 
-struct Minifier {}
-
-impl Minifier {
-    fn is_boolean_attribute(&self, name: &str) -> bool {
-        BOOLEAN_ATTRIBUTES.contains(&name)
-    }
-
-    fn is_default_attribute_value(
-        &self,
-        namespace: Namespace,
-        tag_name: &str,
-        attribute_name: &str,
-        attribute_value: &str,
-    ) -> bool {
-        matches!(
-            (
-                namespace,
-                tag_name,
-                attribute_name,
-                attribute_value.to_ascii_lowercase().trim()
-            ),
-            (Namespace::HTML, "iframe", "height", "150")
-                | (Namespace::HTML, "iframe", "width", "300")
-                | (Namespace::HTML, "iframe", "frameborder", "1")
-                | (Namespace::HTML, "iframe", "loading", "eager")
-                | (Namespace::HTML, "iframe", "fetchpriority", "auto")
-                | (
-                    Namespace::HTML,
-                    "iframe",
-                    "referrerpolicy",
-                    "strict-origin-when-cross-origin"
-                )
-                | (
-                    Namespace::HTML,
-                    "a",
-                    "referrerpolicy",
-                    "strict-origin-when-cross-origin"
-                )
-                | (Namespace::HTML, "a", "target", "_self")
-                | (Namespace::HTML, "area", "target", "_self")
-                | (Namespace::HTML, "area", "shape", "rect")
-                | (
-                    Namespace::HTML,
-                    "area",
-                    "referrerpolicy",
-                    "strict-origin-when-cross-origin"
-                )
-                | (Namespace::HTML, "form", "method", "get")
-                | (Namespace::HTML, "form", "target", "_self")
-                | (Namespace::HTML, "input", "type", "text")
-                | (Namespace::HTML, "input", "size", "20")
-                | (Namespace::HTML, "track", "kind", "subtitles")
-                | (Namespace::HTML, "textarea", "cols", "20")
-                | (Namespace::HTML, "textarea", "rows", "2")
-                | (Namespace::HTML, "textarea", "wrap", "sort")
-                | (Namespace::HTML, "progress", "max", "1")
-                | (Namespace::HTML, "meter", "min", "0")
-                | (Namespace::HTML, "img", "decoding", "auto")
-                | (Namespace::HTML, "img", "fetchpriority", "auto")
-                | (Namespace::HTML, "img", "loading", "eager")
-                | (
-                    Namespace::HTML,
-                    "img",
-                    "referrerpolicy",
-                    "strict-origin-when-cross-origin"
-                )
-                | (Namespace::HTML, "link", "type", "text/css")
-                | (Namespace::HTML, "link", "fetchpriority", "auto")
-                | (
-                    Namespace::HTML,
-                    "link",
-                    "referrerpolicy",
-                    "strict-origin-when-cross-origin"
-                )
-                | (Namespace::HTML, "style", "type", "text/css")
-                | (Namespace::HTML, "script", "type", "text/javascript")
-                | (Namespace::HTML, "script", "type", "text/ecmascript")
-                | (Namespace::HTML, "script", "type", "text/jscript")
-                | (Namespace::HTML, "script", "type", "application/javascript")
-                | (
-                    Namespace::HTML,
-                    "script",
-                    "type",
-                    "application/x-javascript"
-                )
-                | (Namespace::HTML, "script", "type", "application/ecmascript")
-                | (Namespace::HTML, "script", "fetchpriority", "auto")
-                | (
-                    Namespace::HTML,
-                    "script",
-                    "referrerpolicy",
-                    "strict-origin-when-cross-origin"
-                )
-                | (Namespace::HTML, "ol", "type", "1")
-                | (Namespace::HTML, "base", "target", "_self")
-                | (Namespace::HTML, "canvas", "height", "150")
-                | (Namespace::HTML, "canvas", "width", "300")
-        )
-    }
-
-    fn is_conditional_comment(&self, data: &str) -> bool {
-        let trimmed = data.trim();
-
-        if trimmed.starts_with("[if") || trimmed.ends_with("[endif]") {
-            return true;
-        }
-
-        false
-    }
-}
+struct Minifier;
 
 impl VisitMut for Minifier {
     fn visit_mut_element(&mut self, n: &mut Element) {
         n.visit_mut_children_with(self);
 
-        n.children.retain(|child| !matches!(child, Child::Comment(comment) if !self.is_conditional_comment(&comment.data)));
-        n.attributes.retain(|attribute| {
-            if attribute.value.is_none() {
-                return true;
-            }
-
-            match &*attribute.name {
-                _ if self.is_default_attribute_value(
-                    n.namespace,
-                    &n.tag_name,
-                    &attribute.name,
-                    attribute.value.as_ref().unwrap(),
-                ) =>
-                {
-                    false
-                }
-                _ => true,
-            }
-        });
+        steps::remove_comments::remove_comments(n);
+        steps::remove_default_attrs::remove_default_attributes(n);
     }
 
     fn visit_mut_attribute(&mut self, n: &mut Attribute) {
         n.visit_mut_children_with(self);
 
-        match &n.name {
-            name if n.value.is_some() && self.is_boolean_attribute(name) => {
-                n.value = None;
-            }
-            _ => {}
-        }
+        steps::collapse_boolean_attrs::collapse_boolean_attributes(n);
     }
 }
 
 pub fn minify(document: &mut Document) {
-    document.visit_mut_with(&mut Minifier {});
+    document.visit_mut_with(&mut Minifier);
 }
