+++ crates/swc_html_minifier/src/steps/mod.rs
@@ -1,0 +1,3 @@
+pub mod collapse_boolean_attrs;
+pub mod remove_comments;
+pub mod remove_default_attrs;
